refactor(opengemini): move database config to datasource
